refactor(registrations): name the system-req cast, tighten the vote gate
`feedbackDestination` moves next to its only consumer. The token helpers
it shared a file with are genuinely cross-boundary — the follow-up job
signs, the page verifies — but the destination decision is the page's
alone, which is what made that file feel like a grab-bag.

`{ ...req, user: null } as typeof req` becomes `asSystemReq`, beside
`asTrustedReq`. It was doing something subtle enough to need a comment
every time: `overrideAccess` skips collection access but *not*
`filterOptions`, which Payload validates against `req.user`, so a
client-triggered system write touching the events `region` picker was
400ing on a filter the caller could never pass.

The vote gate resolves the event once and applies one condition, instead
of nesting the checks inside `if (eventId != null)`.

**That last one is smaller than I first described it.** I claimed the
gate silently accepted votes when the event couldn't be resolved; it
didn't — a failed lookup was always refused, and only a wholly absent
relationship skipped the gate, which `event: required` makes
unreachable. The mutation check is what caught the overstatement: the
new test passes against the old code too. Kept as a simplification, and
the test kept for its own sake — that a trashed listing stops accepting
the votes deciding whether it stays down was untested.

`eventFeedbackAt` is **not** removed, though it is still write-only.
Generating the migration produced a diff full of main's canonical
columns and no drop at all: every `20260820_*` snapshot dates from
before #642 was rebased onto main, so drizzle diffs from a schema out of
a parallel history. That blocks *any* new migration on this branch, not
just this one, and re-baselining the chain is its own careful change.

Refs #626
